System.Socket.Family.Unix

Description

Documentation

data Unix

The Unix domain socket

data family SocketAddress f

The SocketAddress type is a data family. This allows to provide different data constructors depending on the socket family without knowing all of them in advance or the need to extend this core library.

SocketAddressInet  inetLoopback  8080     :: SocketAddress Inet
SocketAddressInet6 inet6Loopback 8080 0 0 :: SocketAddress Inet6

data family SocketAddress f

The SocketAddress type is a data family. This allows to provide different data constructors depending on the socket family without knowing all of them in advance or the need to extend this core library.

SocketAddressInet  inetLoopback  8080     :: SocketAddress Inet
SocketAddressInet6 inet6Loopback 8080 0 0 :: SocketAddress Inet6

socketAddressUnixPath :: ByteString -> Maybe (SocketAddress Unix)

Creates address which is connected to a filesystem pathname. Returns Nothing if path's length exceeds maximal supported.

socketAddressUnixAbstract :: ByteString -> Maybe (SocketAddress Unix)

Creates address with name in abstract namespace. Returns Nothing if path's length exceeds maximal supported.

getUnixPath :: SocketAddress Unix -> Maybe ByteString

Returns filesystem pathname where address is connected to.

Exceptions

eNoEntry :: SocketException

SocketException "No such file or directory"